requirements.txt
python-dotenv
dspy-ai
streamlit